# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "AmpsHero",
    "description": "AmpsHero API",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users",
        "description": "Get all users with partial filter on username",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user",
        "description": "Create a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/profile/{username}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "User Profile",
        "description": "Display the profile page for a user, identified by their username.",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/create": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "User Creation Form",
        "description": "Render the form page for creating a new user account.",
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get one user data",
        "description": "Get one user data with list of presets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update one user",
        "description": "Update one user (need to be authenticated or admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete one user",
        "description": "Delete one user (need to be authenticated or admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "Get all brands",
        "description": "Get all brands with name and logo",
        "parameters": [
          {
            "name": "page",
            "description": "Page number (default 0)",
            "type": "number",
            "in": "query"
          },
          {
            "name": "limit",
            "description": "Elements per page (default 2)",
            "type": "number",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "summary": "Create a new brand",
        "description": "Create a new brand with name and logo",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "402": {
            "description": "Payment Required"
          }
        }
      }
    },
    "/brands/{id}": {
      "get": {
        "tags": [
          "Brands"
        ],
        "summary": "Get one brand data",
        "description": "Get one brand data",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "summary": "Update one brand",
        "description": "Update one brand",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "logo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "patch": {
        "tags": [
          "Brands"
        ],
        "summary": "Update partially one brand",
        "description": "Update partially one brand",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Brands"
        ],
        "summary": "Delete one brand",
        "description": "Delete one brand",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/amplifiers/": {
      "get": {
        "tags": [
          "Amplifiers"
        ],
        "summary": "Get all amplifiers",
        "description": "Get all amplifiers",
        "parameters": [
          {
            "name": "page",
            "description": "Page number (default 0)",
            "type": "number",
            "in": "query"
          },
          {
            "name": "limit",
            "description": "Elements per page (default 2)",
            "type": "number",
            "in": "query"
          },
          {
            "name": "brand",
            "description": "Brand id",
            "type": "string",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Amplifiers"
        ],
        "summary": "Create a new amplifier",
        "description": "Create a new amplifier",
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/amplifiers/{id}": {
      "get": {
        "tags": [
          "Amplifiers"
        ],
        "summary": "Get one amplifier data",
        "description": "Get one amplifier data",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Amplifiers"
        ],
        "summary": "Update one amplifier data",
        "description": "Update one amplifier data",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Amplifiers"
        ],
        "summary": "Delete one amplifier",
        "description": "Delete one amplifier with its presets",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          }
        }
      }
    },
    "/presets/presets": {
      "get": {
        "tags": [
          "Presets"
        ],
        "summary": "Get all presets",
        "description": "Retrieve a list of all presets, optionally filtering by amplifier and searching by music title. Includes user and amplifier data.",
        "parameters": [
          {
            "name": "page",
            "description": "Page number (default 0)",
            "type": "number",
            "in": "query"
          },
          {
            "name": "limit",
            "description": "Number of elements per page (default 2)",
            "type": "number",
            "in": "query"
          },
          {
            "name": "ampli",
            "description": "Filter by amplifier ID",
            "type": "string",
            "in": "query"
          },
          {
            "name": "musicTitle",
            "description": "Partial search on music title",
            "type": "string",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/presets/": {
      "post": {
        "tags": [
          "Presets"
        ],
        "summary": "Create a new preset",
        "description": "Create a new preset (need to be authenticated)",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/presets(/presets/{id)}": {
      "get": {
        "tags": [
          "Presets"
        ],
        "summary": "Get one preset data",
        "description": "Get one preset data with user and amplifier data",
        "parameters": [
          {
            "name": "id)",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/presets/{id}": {
      "patch": {
        "tags": [
          "Presets"
        ],
        "summary": "Update one preset data",
        "description": "Update one preset data (need to be authenticated or admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "tags": [
          "Presets"
        ],
        "summary": "Delete one preset",
        "description": "Delete one preset (need to be authenticated or admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/uploads/": {
      "post": {
        "tags": [
          "Uploads"
        ],
        "summary": "Upload an image",
        "description": "Upload an image",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}